make test client ext not just work with in mem backend
