+++ core/test-client/src/client_ext.rs
@@ -16,21 +16,17 @@
 
 //! Client extension for tests.
 
-use std::sync::Arc;
 use client::{self, Client};
 use keyring::Keyring;
 use runtime_primitives::{generic::BlockId, StorageMap};
 use runtime::genesismap::{GenesisConfig, additional_storage_with_genesis};
 use executor::NativeExecutor;
+use primitives::{KeccakHasher, RlpCodec};
 use runtime;
 use bft;
-use {Backend, Executor};
 
 /// Extension trait for a test client.
 pub trait TestClient {
-	/// Crates new client instance for tests.
-	fn new_for_tests() -> Self;
-
 	/// Justify and import block to the chain. No finality.
 	fn justify_and_import(&self, origin: client::BlockOrigin, block: runtime::Block) -> client::error::Result<()>;
 
@@ -41,11 +37,11 @@ pub trait TestClient {
 	fn genesis_hash(&self) -> runtime::Hash;
 }
 
-impl TestClient for Client<Backend, Executor, runtime::Block> {
-	fn new_for_tests() -> Self {
-		client::new_with_backend(Arc::new(Backend::new()), NativeExecutor::new(), genesis_storage()).unwrap()
-	}
-
+impl<B, E> TestClient for Client<B, E, runtime::Block>
+    where
+        B: client::backend::Backend<runtime::Block, KeccakHasher, RlpCodec>,
+        E: client::CallExecutor<runtime::Block, KeccakHasher, RlpCodec>
+{
 	fn justify_and_import(&self, origin: client::BlockOrigin, block: runtime::Block) -> client::error::Result<()> {
 		let justification = fake_justify(&block.header);
 		let justified = self.check_justification(block.header, justification)?;
@@ -93,19 +89,4 @@ fn fake_justify(header: &runtime::Header) -> bft::UncheckedJustification<runtime
 		}).collect(),
 		1,
 	)
-}
-
-fn genesis_config() -> GenesisConfig {
-	GenesisConfig::new_simple(vec![
-		Keyring::Alice.to_raw_public().into(),
-		Keyring::Bob.to_raw_public().into(),
-		Keyring::Charlie.to_raw_public().into(),
-	], 1000)
-}
-
-pub fn genesis_storage() -> StorageMap {
-		let mut storage = genesis_config().genesis_map();
-		let block: runtime::Block = client::genesis::construct_genesis_block(&storage);
-		storage.extend(additional_storage_with_genesis(&block));
-		storage
 }
